fix: Organizacion de home en componentes
