+++ app/homec/home.module.css
@@ -1,0 +1,3 @@
+.mainPage{
+  position: relative;
+}
